inventory-backend/requirements.txt
django==3.1.4
psycopg2-binary==2.8.6
djangorestframework==3.12.2
django-cors-headers==3.6.0
django-simple-history==2.12.0
django-extensions==3.1.0
django-cacheops==5.1
django-redis==4.12.1
celery==4.4.6
pyjwt==2.0.0
dateutils==0.6.12
requests==2.22.0
sendgrid==6.4.6
